+++ packages/text-link/src/index.ts
@@ -1,7 +1,7 @@
 export { TextLink } from './Text-Link';
-export { TextLinkButton } from './TextLinkButton';
+export { TextLinkButton } from './Text-Link-Button';
 
 // types
 
 export type { TextLinkProps } from './Text-Link';
-export type { TextLinkButtonProps } from './TextLinkButton';
+export type { TextLinkButtonProps } from './Text-Link-Button';
